sysinfo: wrap GetAllocatableResources if disabled
The resource topology exporter depends on the full
availability of the podresources API, so with *both*
- List
- GetAllocatableResources

However the latter is feature-gated - being alpha,
and in some environments it could not be enabled,
nor the admin would be willing to enable such feature gate.

It is still worthy to be able to *test* and *benchmark*
RTE in these environments, so for *benchmark purposes*
we add a very limited and very intentionally narrow
emulation of GetAllocatableResources functionality,
reading raw data from the system the same way the kubelet would.

Of course this is intentionally limited and will not
be evolved (barring bugs and extremely simple additions)
because the recommended and supported way is to
just enable the full feature set of the podresources API.

Signed-off-by: Francesco Romani <fromani@redhat.com>