Add release lane
From now on, for every new tag we'll build and push an image to quay based on that tag.
There is no need with testing the image since the openshift CI covers that for us.
We can also trigger the building manually from the Actions tab.

This workflow inspired by the work that was done in: https://github.com/k8stopologyawareschedwg/deployer/blob/main/.github/workflows/release.yml

Signed-off-by: Talor Itzhak <titzhak@redhat.com>